Finished section 3 of Treehouse course.
Utilising fragments, I now have a two-tabbed application which shows an Inbox and Friends tab. Both tabs currently show an empty list view with an empty message from a TextView.

Couple of issues to do with Eclipse vs. Android Studio but by the end of the section, this was resolved.